Default to non EFI and add option to enable it

+++ makeUSB.sh
@@ -7,7 +7,9 @@ set -o nounset
 
 # Defaults
 scriptname=$(basename "$0")
+eficonfig=0
 interactive=0
+data_part=2
 data_fmt="vfat"
 efi_mnt="/mnt/MBU-EFI/"
 data_mnt="/mnt/MBU-DATA/"
@@ -21,6 +23,7 @@ showUsage() {
 
 	 device                         Device to modify (e.g. /dev/sdb)
 	 fs-type                        Filesystem type for the data partition [ext3|ext4|vfat|ntfs]
+	  -e,  --efi                    Enable EFI compatibility
 	  -i,  --interactive            Launch gdisk to create a hybrid MBR
 	  -l,  --log                    Save debug information to log
 	  -h,  --help                   Display this message
@@ -50,6 +53,11 @@ while [ "$#" -gt 0 ]; do
 		-h|--help)
 			showUsage
 			exit 0
+			;;
+		-e|--efi)
+			eficonfig=1
+			data_part=3
+			shift
 			;;
 		-i|--interactive)
 			interactive=1
@@ -149,14 +157,16 @@ else
 	cleanUp 10
 fi
 
-# Create EFI System partition (50M)
-printf 'Creating EFI System partition on %s... ' "$usb_dev"
-if sgdisk --new 2::+50M --typecode 2:ef00 --change-name 2:"EFI System" \
-    "$usb_dev" >> "$log_file" 2>&1; then
-	printf 'OK\n'
-else
-	printf 'FAILED\n'
-	cleanUp 10
+if [ "$eficonfig" -eq 1 ]; then
+	# Create EFI System partition (50M)
+	printf 'Creating EFI System partition on %s... ' "$usb_dev"
+	if sgdisk --new 2::+50M --typecode 2:ef00 --change-name 2:"EFI System" \
+	    "$usb_dev" >> "$log_file" 2>&1; then
+		printf 'OK\n'
+	else
+		printf 'FAILED\n'
+		cleanUp 10
+	fi
 fi
 
 # Create data partition
@@ -176,8 +186,8 @@ case "$data_fmt" in
 		cleanUp 1
 		;;
 esac
-if sgdisk --new 3::: --typecode 3:"$type_code" --change-name 3:"$part_name" \
-    "$usb_dev" >> "$log_file" 2>&1; then
+if sgdisk --new ${data_part}::: --typecode ${data_part}:"$type_code" \
+    --change-name ${data_part}:"$part_name" "$usb_dev" >> "$log_file" 2>&1; then
 	printf 'OK\n'
 else
 	printf 'FAILED\n'
@@ -190,11 +200,20 @@ umount --force ${usb_dev}* 2>/dev/null
 # Create hybrid MBR
 if [ "$interactive" -eq 0 ]; then
 	printf 'Creating hybrid MBR on %s... ' "${usb_dev}"
-	if sgdisk --hybrid 1:2:3 "$usb_dev" >> "$log_file" 2>&1; then
-		printf 'OK\n'
+	if [ "$eficonfig" -eq 1 ]; then
+		if sgdisk --hybrid 1:2:3 "$usb_dev" >> "$log_file" 2>&1; then
+			printf 'OK\n'
+		else
+			printf 'FAILED\n'
+			cleanUp 10
+		fi
 	else
-		printf 'FAILED\n'
-		cleanUp 10
+		if sgdisk --hybrid 1:2 "$usb_dev" >> "$log_file" 2>&1; then
+			printf 'OK\n'
+		else
+			printf 'FAILED\n'
+			cleanUp 10
+		fi
 	fi
 else
 	# Create hybrid MBR manually
@@ -203,8 +222,8 @@ else
 fi
 
 # Set bootable flag for data partion
-printf 'Setting bootable flag on %s... ' "${usb_dev}3"
-if sgdisk --attributes 3:set:2 "$usb_dev" >> "$log_file" 2>&1; then
+printf 'Setting bootable flag on %s... ' "${usb_dev}${data_part}"
+if sgdisk --attributes ${data_part}:set:2 "$usb_dev" >> "$log_file" 2>&1; then
 	printf 'OK\n'
 else
 	printf 'FAILED\n'
@@ -223,18 +242,21 @@ else
 	cleanUp 10
 fi
 
-# Format EFI System partition
-printf 'Formatting EFI System partition on %s... ' "${usb_dev}2"
-if mkfs.vfat -v -F 32 "${usb_dev}2" >> "$log_file" 2>&1; then
-	printf 'OK\n'
-else
-	printf 'FAILED\n'
-	cleanUp 10
+if [ "$eficonfig" -eq 1 ]; then
+	# Format EFI System partition
+	printf 'Formatting EFI System partition on %s... ' "${usb_dev}2"
+	if mkfs.vfat -v -F 32 "${usb_dev}2" >> "$log_file" 2>&1; then
+		printf 'OK\n'
+	else
+		printf 'FAILED\n'
+		cleanUp 10
+	fi
 fi
 
 # Format data partition
-printf 'Formatting data partition as %s on %s... ' "$data_fmt" "${usb_dev}3"
-if mkfs.${data_fmt} "${usb_dev}3" >> "$log_file" 2>&1; then
+printf 'Formatting data partition as %s on %s... ' \
+    "$data_fmt" "${usb_dev}${data_part}"
+if mkfs.${data_fmt} "${usb_dev}${data_part}" >> "$log_file" 2>&1; then
 	printf 'OK\n'
 else
 	printf 'FAILED\n'
@@ -244,42 +266,47 @@ fi
 # Unmount device
 umount --force ${usb_dev}* 2>/dev/null
 
-# Mount EFI System partition
-printf 'Mounting EFI System partition on %s... ' "$efi_mnt"
-if mkdir -p "$efi_mnt" \
-    && mount -v "${usb_dev}2" "$efi_mnt" >> "$log_file" 2>&1; then
-	printf 'OK\n'
-else
-	printf 'FAILED\n'
-	cleanUp 10
+if [ "$eficonfig" -eq 1 ]; then
+	# Mount EFI System partition
+	printf 'Mounting EFI System partition on %s... ' "$efi_mnt"
+	if mkdir -p "$efi_mnt" \
+	    && mount -v "${usb_dev}2" "$efi_mnt" >> "$log_file" 2>&1; then
+		printf 'OK\n'
+	else
+		printf 'FAILED\n'
+		cleanUp 10
+	fi
 fi
 
 # Mount data partition
 printf 'Mounting data partition on %s... ' "$data_mnt"
 if mkdir -p "$data_mnt" \
-    && mount -v "${usb_dev}3" "$data_mnt" >> "$log_file" 2>&1; then
+    && mount -v "${usb_dev}${data_part}" "$data_mnt" >> "$log_file" 2>&1; then
 	printf 'OK\n'
 else
 	printf 'FAILED\n'
 	cleanUp 10
 fi
 
-# Install GRUB for EFI
-printf 'Installing GRUB for EFI on %s... ' "$usb_dev"
-if $grub_cmd --target=x86_64-efi --efi-directory=$efi_mnt \
-    --boot-directory=${data_mnt}boot --removable --recheck >> "$log_file" 2>&1; then
-	printf 'OK\n'
-else
-	printf 'FAILED\n'
-fi
+if [ "$eficonfig" -eq 1 ]; then
+	# Install GRUB for EFI
+	printf 'Installing GRUB for EFI on %s... ' "$usb_dev"
+	if $grub_cmd --target=x86_64-efi --efi-directory=$efi_mnt \
+	    --boot-directory=${data_mnt}boot --removable --recheck >> "$log_file" 2>&1; then
+		printf 'OK\n'
+	else
+		printf 'FAILED\n'
+		cleanUp 10
+	fi
 
-# Unmount EFI System partition
-printf 'Unmounting EFI System partition on %s... ' "$efi_mnt"
-if umount -v $efi_mnt >> "$log_file" 2>&1; then
-	printf 'OK\n'
-else
-	printf 'FAILED\n'
-	cleanUp 10
+	# Unmount EFI System partition
+	printf 'Unmounting EFI System partition on %s... ' "$efi_mnt"
+	if umount -v $efi_mnt >> "$log_file" 2>&1; then
+		printf 'OK\n'
+	else
+		printf 'FAILED\n'
+		cleanUp 10
+	fi
 fi
 
 # Install GRUB for BIOS
@@ -293,10 +320,10 @@ else
 fi
 
 # Install fallback GRUB
-printf 'Installing fallback GRUB on %s... ' "${usb_dev}3"
+printf 'Installing fallback GRUB on %s... ' "${usb_dev}${data_part}"
 if $grub_cmd --force --target=i386-pc \
     --boot-directory=${data_mnt}boot \
-    --recheck "${usb_dev}3" >> "$log_file" 2>&1; then
+    --recheck "${usb_dev}${data_part}" >> "$log_file" 2>&1; then
 	printf 'OK\n'
 else
 	printf 'FAILED\n'
